requirements.txt
streamlit==0.88.0
plotly==5.3.1
matplotlib==3.4.3
seaborn==0.11.2
pandas==1.3.3
numpy==1.21.2
Pillow==8.3.2
scikit_learn==0.24.2
